backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
open3d>=0.17.0
trimesh>=3.21.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
httpx>=0.24.0
psutil>=5.9.0